requirements-grpc.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
googleapis-common-protos==1.62.0

# gRPC health checking
grpc-health-checking==1.0.0

# gRPC reflection
grpc-reflection==0.1.0

# gRPC web
grpc-interceptor==0.15.4
grpc-stubs==1.24.7

# Async gRPC
grpclib==0.4.7
h2==4.1.0

# Protocol Buffers
protobuf-stubs==1.0.0
mypy-protobuf==3.4.0

# gRPC testing
grpcio-testing==1.60.0
